+++ Projects/Axiom/Source/RenderSystems/OpenGL/GLHardwareIndexBuffer.cs
@@ -154,7 +154,7 @@ namespace Axiom.RenderSystems.OpenGL
 
 			isLocked = true;
 
-			return new IntPtr( ptr.ToInt32() + offset );
+            return new IntPtr(ptr.ToInt64() + offset);
 		}
 
 		/// <summary>
